Improves movie loading and UI state management
Refactors the HomeScreen to use a sealed MoviesUiState for managing loading, success, and error states.
Implements pagination using LazyColumn and handles loading more movies when the user scrolls to the end of the list, using collectAsStateWithLifecycle.
Moves the API key to a constant within the MovieRepositoryImpl for better security and maintainability.
Updates the OnboardingScreen UI with background colors, text styles, and button styling.